employee_management/requirements.txt
-e git+https://github.com/SanjeevKumar-postiefs/employee-project.git@6781711e5040a2c362e8a8a8321dc3cddd5c8ed8#egg=activity_tracker&subdirectory=desktop_agent
altgraph==0.17.4
annotated-types==0.7.0
anyascii==0.3.2
anyio==4.8.0
asgiref==3.8.1
better-profanity==0.7.0
blis==1.2.0
catalogue==2.0.10
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.20.0
colorama==0.4.6
confection==0.1.5
contractions==0.1.73
cymem==2.0.11
distro==1.9.0
Django==4.2
django-cors-headers==4.7.0
djangorestframework==3.15.2
dnspython==2.7.0
email_validator==2.2.0
-e git+https://github.com/SanjeevKumar-postiefs/employee-project.git@6781711e5040a2c362e8a8a8321dc3cddd5c8ed8#egg=employee_activity_tracker&subdirectory=desktop_agent
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
filelock==3.17.0
fsspec==2025.2.0
gingerit==0.0.0.1
gpt4all==2.8.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.1
idna==3.10
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
langcodes==3.5.0
language_data==1.3.0
language_tool_python==2.9.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
murmurhash==1.0.12
mysqlclient==2.2.4
networkx==3.4.2
nltk==3.9.1
numpy==2.2.3
openai==1.64.0
packaging==24.2
pefile==2023.2.7
pillow==11.1.0
preshed==3.0.9
psutil==7.0.0
pyahocorasick==2.1.0
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
pyspellchecker==0.8.2
python-dotenv==1.0.1
pytz==2024.2
pywin32==308
pywin32-ctypes==0.2.3
PyYAML==6.0.2
redis==5.2.0
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.5.3
setuptools==75.8.2
shellingham==1.5.4
smart-open==7.1.0
sniffio==1.3.1
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlitedict==2.1.0
sqlparse==0.5.1
srsly==2.5.1
sympy==1.13.1
textsearch==0.0.24
thinc==8.3.4
tiktoken==0.9.0
tokenizers==0.21.0
toml==0.10.2
torch==2.6.0
tqdm==4.67.1
transformers==4.49.0
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
wasabi==1.1.3
watchdog==6.0.0
weasel==0.4.1
wrapt==1.17.2
